[releng] Switch to Sirius Web 2026.3.2

+++ backend/application/syson-sysml-import/src/main/java/org/eclipse/syson/sysml/SysmlToAst.java
@@ -105,7 +105,7 @@ public class SysmlToAst {
             if (sysmlInputPath != null) {
                 sysmlInputPath.toFile().delete();
             }
-            if (this.cliPath == null) {
+            if (this.cliPath == null && sysIdeInputPath != null) {
                 sysIdeInputPath.toFile().delete();
             }
         }
